# api/openapi.yaml
openapi: 3.1.0
info:
  title: dra-workflow API
  version: 0.1.0
  summary: Workflow-driven, AI-assisted IT service disaster recovery planning and execution.
  description: |
    REST API of the dra-workflow backend (Rust). It implements the data model in
    `docs/domain/dr-plan-model.md` and the workflows in `docs/workflows/`.

    **Hierarchy:** Tenant → IT service → microservice → DR items.

    **Methodology:** NIST SP 800-34 Rev. 1 and BSI-Standard 200-4 (see `docs/domain/standards-mapping.md`).

    ## Conventions
    - **Tenant from the token:** the tenant is **not** part of the path. The backend reads it from the
      `tenant_id` claim of the JWT access token and scopes every query to it. A token without a valid
      tenant claim, or for a tenant the user is not a member of, gets `403`. Ids belonging to another
      tenant return `404`. Only `GET/POST /tenants` (membership list, tenant creation) are tenant-independent.
    - **Collections** are nested under their parent (for example `/services/{serviceId}/microservices`).
      **Single items** are addressed flat by id (for example `/microservices/{microserviceId}`).
    - IDs are UUIDs. Durations are **integer minutes**. Timestamps are RFC 3339 UTC.
    - Updates use `PATCH` with `application/merge-patch+json` (RFC 7396).
    - Optimistic concurrency: every resource has a `version`. Reads return an `ETag`, and `PATCH`/`DELETE`
      accept `If-Match`. A stale version returns `412`.
...
      with `issues`.
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /services/{serviceId}/action-items:
    parameters:
      - $ref: '#/components/parameters/ServiceId'
    get:
      tags: [Action Items]
      operationId: listActionItems
      summary: List improvement actions (workflow step 15)
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/ActionItemStatus' }
        - name: source
          in: query
          schema: { $ref: '#/components/schemas/ActionItemSource' }
      responses:
        '200':
          description: Action items
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ActionItem' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [Action Items]
      operationId: createActionItem
      summary: Create an action item
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ActionItemCreate' }
      responses:
        '201':
          description: Created
          headers:
            Location: { $ref: '#/components/headers/Location' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionItem' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
  /action-items/{actionItemId}:
    parameters:
      - $ref: '#/components/parameters/ActionItemId'
    patch:
      tags: [Action Items]
      operationId: updateActionItem
      summary: Update an action item
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: '#/components/schemas/ActionItemFields' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ActionItem' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }

  # ───────────────────────────── AI Assistance ─────────────────────────────
  /ai/suggestions:
    get:
      tags: [AI Assistance]
      operationId: listAiSuggestions
      summary: List AI suggestions
      parameters:
        - name: serviceId
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/AiSuggestionStatus' }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Suggestions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiSuggestionPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
    post:
      tags: [AI Assistance]
      operationId: requestAiSuggestion
      summary: Ask the AI assistant for a suggestion
      description: |
        Returns `202` with the suggestion (currently computed synchronously, so the status is already `proposed`;
        clients should still handle `pending` and poll the `Location`). The context sent to the provider is minimized:
        person contact details are removed and secrets are never part of a plan. Returns `503` while no AI
        provider is configured or AI is disabled for the tenant.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AiSuggestionRequest' }
      responses:
        '202':
          description: Accepted
          headers:
            Location: { $ref: '#/components/headers/Location' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiSuggestion' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
  /ai/suggestions/{suggestionId}:
    parameters:
      - $ref: '#/components/parameters/SuggestionId'
    get:
      tags: [AI Assistance]
      operationId: getAiSuggestion
      summary: Get a suggestion
      responses:
        '200':
          description: Suggestion
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiSuggestion' }
        '404': { $ref: '#/components/responses/NotFound' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }
  /ai/suggestions/{suggestionId}/decision:
    parameters:
      - $ref: '#/components/parameters/SuggestionId'
    post:
      tags: [AI Assistance]
      operationId: decideAiSuggestion
      summary: Accept, edit or reject proposals
      description: |
        Accepted or edited proposals are applied as normal create/update operations, with `origin: ai_accepted`,
        and go through the same validation. This is the only way AI output reaches the plan.
        **Not implemented yet:** `accept`/`edit` return `501`; rejecting is supported.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AiDecision' }
      responses:
        '200':
          description: Decision applied
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiDecisionResult' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Error' }

  # ───────────────────────────── Recovery Runs ─────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC access token (JWT). Required claims: `sub` (user) and `tenant_id` (UUID of the active tenant).
        The tenant role may come from the token or from the tenant membership. The identity provider is
        still open (docs/overview/open-questions.md).
  headers:
    Location:
      description: URL of the created resource
      schema: { type: string }
...
  parameters:
    ServiceId:
      name: serviceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ActionItemId:
      name: actionItemId
      in: path
      required: true
      schema: { type: string, format: uuid }
    SuggestionId:
      name: suggestionId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from a previous page's `nextCursor`
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Error:
      description: Unexpected error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found (or it belongs to a different tenant than the token's)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    Conflict:
      description: Conflicts with the current state (for example the resource is still referenced, or the transition is invalid)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    ValidationFailed:
      description: Input validation, integrity rule or workflow gate failed
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ValidationProblem' }
    TooManyRequests:
      description: Rate limit exceeded (AI requests)
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ServiceUnavailable:
      description: The AI provider is unavailable. All non-AI functions keep working.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details
      properties:
        type: { type: string, format: uri-reference, default: about:blank }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
    ValidationProblem:
      allOf:
        - $ref: '#/components/schemas/Problem'
        - type: object
          properties:
            issues:
              type: array
              items: { $ref: '#/components/schemas/Issue' }
    Issue:
      type: object
      description: A validation error, integrity-rule violation or gate failure
      required: [ruleId, severity, message]
      properties:
        ruleId:
          type: string
          description: Stable rule id, for example `RTO_EXCEEDS_SERVICE_RTO`, `DEPENDENCY_RTO_CONFLICT`, `STEP_WITHOUT_OWNER`, `STRATEGY_GAP`, `ROLE_WITHOUT_DEPUTY`
          examples: [DEPENDENCY_RTO_CONFLICT]
        severity: { type: string, enum: [blocking, warning, info] }
        message: { type: string }
        field: { type: string, description: 'JSON pointer to the offending input field, if any' }
        entityType: { type: string }
        entityId: { type: string, format: uuid }
        workflowStep: { $ref: '#/components/schemas/WorkflowStepKey' }
        standardRef:
          type: string
          description: Related standard reference
          examples: ['NIST SP 800-34 §3.2', 'BSI 200-4 Soll-Ist-Vergleich']
    ResourceMeta:
      type: object
      required: [id, tenantId, version, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid, readOnly: true }
        tenantId: { type: string, format: uuid, readOnly: true }
        version: { type: integer, readOnly: true, description: Incremented on every change (optimistic concurrency) }
        createdAt: { type: string, format: date-time, readOnly: true }
        createdBy: { type: string, readOnly: true }
        updatedAt: { type: string, format: date-time, readOnly: true }
        updatedBy: { type: string, readOnly: true }
    ContentMeta:
      type: object
      properties:
        origin:
          type: string
          enum: [user, ai_accepted, import]
          readOnly: true
          description: Provenance of the content
        aiSuggestionId:
          type: [string, 'null']
          format: uuid
          readOnly: true
    PageInfo:
      type: object
      properties:
        nextCursor: { type: [string, 'null'] }
...
    WorkflowStepKey:
      type: string
      enum:
        - define_service
        - map_dependencies
        - business_impact
        - brainstorm_scenarios
        - consolidate_scenarios
        - select_scenarios
        - recovery_objectives
        - recovery_strategies
        - runbooks
        - roles
        - communication
        - review_approve
        - test
        - measure
        - improve
...
    ActionItemSource:
      type: string
      enum: [test, run, review, ai_gap_check, manual]
    ActionItemStatus:
      type: string
      enum: [open, in_progress, done, wont_fix]
    AiSuggestionKind:
      type: string
      description: What the assistant should do. Each kind maps to a workflow step or a recovery-run situation.
      enum:
        - describe_service
        - suggest_microservices
        - suggest_dependencies
        - check_bia
        - suggest_scenarios
        - consolidate_scenarios
        - assess_scenarios
        - suggest_objectives
        - suggest_strategies
        - gap_check
        - draft_runbook
        - review_runbook
        - check_roles
        - draft_communication
        - review_plan
        - tabletop_script
        - summarize_test
        - lessons_learned
        - troubleshoot_step
        - status_update
    AiSuggestionStatus:
      type: string
      enum: [pending, proposed, accepted, partially_accepted, rejected, failed]
...
    ActionItemFields:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        ownerPersonId: { type: [string, 'null'], format: uuid }
        dueDate: { type: [string, 'null'], format: date }
        status: { $ref: '#/components/schemas/ActionItemStatus' }
        relatedEntity:
          type: [object, 'null']
          properties:
            type: { type: string, examples: [runbook_step] }
            id: { type: string, format: uuid }
    ActionItemCreate:
      allOf:
        - $ref: '#/components/schemas/ActionItemFields'
        - required: [title]
    ActionItem:
      allOf:
        - $ref: '#/components/schemas/ResourceMeta'
        - $ref: '#/components/schemas/ContentMeta'
        - $ref: '#/components/schemas/ActionItemFields'
        - type: object
          required: [title, status, source]
          properties:
            serviceId: { type: string, format: uuid, readOnly: true }
            source: { $ref: '#/components/schemas/ActionItemSource' }
            sourceId: { type: [string, 'null'], format: uuid, readOnly: true }

    # ─────────── AI ───────────
    AiSuggestionRequest:
      type: object
      required: [kind, serviceId]
      properties:
        kind: { $ref: '#/components/schemas/AiSuggestionKind' }
        serviceId: { type: string, format: uuid }
        microserviceId: { type: string, format: uuid }
        scenarioId: { type: string, format: uuid }
        runbookId: { type: string, format: uuid }
        recoveryRunId: { type: string, format: uuid }
        runbookStepId: { type: string, format: uuid }
        prompt: { type: string, maxLength: 4000, description: 'Optional free text, for example an error message during recovery' }
        language: { type: string, enum: [en, de], default: en }
    AiProposal:
      type: object
      required: [proposalId, entityType, operation, data]
      properties:
        proposalId: { type: string }
        entityType:
          type: string
          examples: [scenario, dependency, recovery_strategy, runbook_step, communication_rule]
        operation: { type: string, enum: [create, update, merge, advice] }
        targetId: { type: [string, 'null'], format: uuid, description: For update or merge }
        data:
          type: object
          description: Payload shaped like the corresponding Create or Fields schema (or `{ text }` for advice)
        rationale: { type: string }
        confidence: { type: string, enum: [low, medium, high] }
    AiSuggestion:
      type: object
      required: [id, kind, status]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: '#/components/schemas/AiSuggestionKind' }
        status: { $ref: '#/components/schemas/AiSuggestionStatus' }
        serviceId: { type: string, format: uuid }
        context: { $ref: '#/components/schemas/AiSuggestionRequest' }
        proposals:
          type: array
          items: { $ref: '#/components/schemas/AiProposal' }
        summary: { type: string }
        model: { type: string }
        error: { type: [string, 'null'] }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        decidedBy: { type: [string, 'null'] }
        decidedAt: { type: [string, 'null'], format: date-time }
    AiSuggestionPage:
      allOf:
        - $ref: '#/components/schemas/PageInfo'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/AiSuggestion' }
    AiDecision:
      type: object
      required: [decisions]
      properties:
        decisions:
          type: array
          minItems: 1
          items:
            type: object
            required: [proposalId, decision]
            properties:
              proposalId: { type: string }
              decision: { type: string, enum: [accept, edit, reject] }
              editedData: { type: object, description: Required for `edit` }
        comment: { type: string }
    AiDecisionResult:
      type: object
      properties:
        suggestion: { $ref: '#/components/schemas/AiSuggestion' }
        applied:
          type: array
          items:
            type: object
            properties:
              proposalId: { type: string }
              entityType: { type: string }
              entityId: { type: string, format: uuid }
              operation: { type: string, enum: [create, update, merge] }

    # ─────────── Recovery Runs ───────────
...